Fixed race condition in GetTypeInfo method
It was observed that a new ConvertTypeInfo instance was created twice for two threads simultaneously, adding the same type name to the assembly, which caused an ArgumentException "Duplicate type name within an assembly". This is the normal pattern when using a ConcurrentDictionary.

To resolve this, it is ensured that the new instance (emitting that type) is only created once. A normal Dictionary and locks are used now.